The plans changed!
I realized that the way I want to solve this ploblem is too complex and used OOP way too much! To keep it simple, I added all the functonality needed in the board module in order  it can get imported in another project  and run just fine. Also, the board module now contains the board.py and the __init__.py file. FInally, I desided to extend the functionality of my board module, (so it runs chess), in the main.py file without using OOP. The next goal is to keep updating the main.py file and clean up the board.py file