changes to board.py, main.py
The board module is now complete. I fixed a visual bug in the numbers and letters of the board. Then, I removed all the __init__ arguments and in order to modify soem values you do it after the object is declared. In addition, I used encapsulation;  I made the size and the square_size (now __size and __square_size) methods private to prevent someone updating one of the two and not both methods. If someone needs to update the size of the chess board he 'll use the size() function that is the setter of the __size method and updates the __square_size variable. Now This is the last update of the main branch. Now I 'll be working on the chess-board branch to add more chess features, such as piese generation, movement etc.